# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "fluentlo API",
        "description": "Laravel 12 + Sanctum + Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/app/version/check": {
            "post": {
                "tags": [
                    "App Version"
                ],
                "summary": "Check for app updates (Public)",
                "operationId": "app_check_version",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "platform",
                                    "version",
                                    "build_number"
                                ],
                                "properties": {
                                    "platform": {
                                        "type": "string",
                                        "enum": [
                                            "ios",
                                            "android"
                                        ],
                                        "example": "ios"
                                    },
                                    "version": {
                                        "type": "string",
                                        "example": "1.0.0"
                                    },
                                    "build_number": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Version check result"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/v1/app/version/latest": {
            "get": {
                "tags": [
                    "App Version"
                ],
                "summary": "Get latest version info (Public)",
                "operationId": "app_latest_version",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "query",
                        "description": "Platform (ios/android)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Latest version info"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/v1/admin/app-versions": {
            "get": {
                "tags": [
                    "App Version"
                ],
                "summary": "List app versions (Admin)",
                "operationId": "admin_app_versions_index",
                "parameters": [
                    {
                        "name": "platform",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ios",
                                "android"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App versions list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "App Version"
                ],
                "summary": "Create app version (Admin)",
                "operationId": "admin_app_versions_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "platform",
                                    "version",
                                    "build_number"
                                ],
                                "properties": {
                                    "platform": {
                                        "type": "string",
                                        "enum": [
                                            "ios",
                                            "android"
                                        ]
                                    },
                                    "version": {
                                        "type": "string",
                                        "example": "1.0.1"
                                    },
                                    "build_number": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "force_update": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Update Available",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "<p>New features</p>",
                                        "nullable": true
                                    },
                                    "store_links": {
                                        "properties": {
                                            "app_store": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "google_play": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "direct": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "App version created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/v1/admin/app-versions/{appVersion}/toggle-status": {
            "post": {
                "tags": [
                    "App Version"
                ],
                "summary": "Toggle app version status (Admin)",
                "operationId": "admin_app_versions_toggle_status",
                "parameters": [
                    {
                        "name": "appVersion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status toggled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}